feat: remove self-managed plugin update checks, rely on marketplace
